CustomGroup.get - add pseudoconstant field handling
