+++ Civi/Api4/Generic/BasicGetAction.php
@@ -13,7 +13,6 @@
 namespace Civi\Api4\Generic;
 
 use Civi\API\Exception\NotImplementedException;
-use Civi\Api4\Utils\FormattingUtil;
 
 /**
  * Retrieve $ENTITIES based on criteria specified in the `where` parameter.
@@ -22,6 +21,7 @@ use Civi\Api4\Utils\FormattingUtil;
  */
 class BasicGetAction extends AbstractGetAction {
   use Traits\ArrayQueryActionTrait;
+  use Traits\PseudoconstantOutputTrait;
 
   /**
    * @var callable
@@ -92,32 +92,6 @@ class BasicGetAction extends AbstractGetAction {
       return call_user_func($this->getter, $this);
     }
     throw new NotImplementedException('Getter function not found for api4 ' . $this->getEntityName() . '::' . $this->getActionName());
-  }
-
-  /**
-   * Evaluate :pseudoconstant suffix expressions and replace raw values with option values
-   *
-   * @param $records
-   * @throws \CRM_Core_Exception
-   */
-  protected function formatRawValues(&$records) {
-    // Pad $records and $fields with pseudofields
-    $fields = $this->entityFields();
-    foreach ($records as &$values) {
-      foreach ($this->entityFields() as $field) {
-        $values += [$field['name'] => $field['default_value'] ?? NULL];
-        if (!empty($field['options'])) {
-          foreach ($field['suffixes'] ?? array_keys(\CRM_Core_SelectValues::optionAttributes()) as $suffix) {
-            $pseudofield = $field['name'] . ':' . $suffix;
-            if (!isset($values[$pseudofield]) && isset($values[$field['name']]) && $this->_isFieldSelected($pseudofield)) {
-              $values[$pseudofield] = $values[$field['name']];
-            }
-          }
-        }
-      }
-      // Swap raw values with pseudoconstants
-      FormattingUtil::formatOutputValues($values, $fields, $this->getActionName());
-    }
   }
 
 }
